requirements.txt
# Exact versions from att_map conda environment (verified working)
torch==2.10.0
torchvision==0.25.0
diffusers==0.21.2
transformers==4.30.2
accelerate==0.23.0
daam==0.2.0
huggingface_hub==0.36.2
safetensors==0.7.0
tokenizers==0.13.3
Pillow==12.1.1
matplotlib==3.10.8
numpy==2.2.6
scipy==1.15.3
ftfy==6.3.1
regex==2026.1.15
requests==2.32.5
filelock==3.25.2
tqdm==4.67.3
